Fixed an issue with deleting a category

+++ routes/api/category-routes.js
@@ -67,7 +67,21 @@ router.put('/:id', async (req, res) => {
 
 router.delete('/:id', async (req, res) => {
   // delete a category by its `id` value
+
   try {
+
+    const productData = await Product.destroy({
+      where: {
+        category_id: req.params.id,
+      },
+    });
+
+    if (!productData)
+    {
+      res.status(404).json(categoryData);
+      return;
+    }
+
     const categoryData = await Category.destroy({
       where: {
         id: req.params.id,
